+++ spec/color_ls/flags_spec.rb
